requirements.txt
sounddevice
numpy
scipy
flask
fastapi
uvicorn
httpx
web3
eth-account
python-dotenv
x402
